+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.3.5)
 
+- Add `includeMembersRegex` to `GroupTraits`. This can be used to filter group members by id/name using a regular expression.
 - [The next improvement]
 
 #### 8.3.4 - 2023-09-15